Logic to allow Gad to steal scrap from Jerrod

+++ Assets/Production/0_Code/HumanBuilders/Subsystems/VSave/Variable.cs
@@ -75,7 +75,7 @@ namespace HumanBuilders {
             BoolValue = value;
             break;
           case VariableType.Integer:
-            IntegerValue = value;
+            IntegerValue = (int)value;
             break;
           case VariableType.Float:
             FloatValue = value;
